Safari-611.1.10/LayoutTests/editing/pasteboard/datatransfer-types-dropping-text-file-promise-expected.txt

When dropping a file promise, dataTransfer.types must contain "Files" and not "text/uri-list". This test requires eventSender.beginDragWithFilePromises.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS dataTransfer.types.includes("Files") is true
PASS dataTransfer.types.includes("text/uri-list") is false
PASS dataTransfer.getData("url") is ""
PASS successfullyParsed is true

TEST COMPLETE
